update and move validate_aws_env_vars()
